requirements.txt
amqp==5.1.1
asgiref==3.7.1
async-timeout==4.0.2
attrs==23.1.0
billiard==3.6.4.0
celery==5.2.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cron-descriptor==1.4.0
Django==4.2.1
django-celery-beat==2.5.0
django-debug-toolbar==4.1.0
django-environ==0.10.0
django-filter==23.2
django-timezone-field==5.0
djangorestframework==3.14.0
drf-spectacular==0.26.2
exceptiongroup==1.1.1
factory-boy==3.2.1
Faker==18.9.0
geographiclib==2.0
geopy==2.3.0
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.17.3
kombu==5.2.4
numpy==1.24.3
packaging==23.1
pandas==2.0.1
pluggy==1.0.0
prompt-toolkit==3.0.38
psycopg2==2.9.6
pyrsistent==0.19.3
pytest==7.3.1
pytest-django==4.5.2
python-crontab==2.7.1
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
redis==4.5.5
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.6.2
tzdata==2023.3
uritemplate==4.1.1
vine==5.0.0
wcwidth==0.2.6
